Move product home into crashstats app
This moves the product home view into the crashstats app and gets rid of the
home app. Having a home app is weird. The crashstats app covers so much of
the basic functionality of crash stats that it makes sense to cover this, too.

Further, we're going to put the home page in the crashstats app which will
list all the products. Products will eventually get moved to this app, too.
