+++ webapp-django/crashstats/base/jinja2/crashstats_base.html
@@ -87,8 +87,8 @@
                     <optgroup label=" ">
                         <option
                             value="products"
-                            data-url-product="{{ url('home:product_home', '__PRODUCT__') }}"
-                            data-url-version="{{ url('home:product_home', '__PRODUCT__') }}?version=__VERSION__"
+                            data-url-product="{{ url('crashstats:product_home', '__PRODUCT__') }}"
+                            data-url-version="{{ url('crashstats:product_home', '__PRODUCT__') }}?version=__VERSION__"
                             {% if report == 'products' %}selected{% endif %}
                         >Overview</option>
                     </optgroup>
